requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
da-vinci==0.4.0
Django==5.0.8
django-admin-thumbnails==0.2.9
django-allauth==64.0.0
django-thumbnails==0.8.0
et-xmlfile==1.1.0
fonttools==4.53.1
git-filter-repo==2.45.0
gunicorn==23.0.0
html5lib==1.1
idna==3.7
lxml==5.3.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
razorpay==1.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
